+++ tests/tests/src/core/Core_TempTests.mjs
@@ -249,6 +249,10 @@ let x = Symbol.for("Foo");
 
 console.log(x);
 
+let array$1 = "foo"[Symbol.iterator]().toArray();
+
+console.log(array$1);
+
 console.info("");
 
 console.info("Global namespace");
@@ -334,7 +338,6 @@ export {
   _formatter,
   formatter,
   segments,
-  array,
   date,
   dict,
   dict2,
@@ -351,6 +354,7 @@ export {
   set,
   regexp,
   x,
+  array$1 as array,
   timeout,
   z,
   intFromBigInt,