+++ yarn.config.cjs
@@ -14,7 +14,7 @@ const execPromise = util.promisify(exec);
  * @param {Yarn.Constraints.Context} ctx
  */
 async function enforceCompilerMeta({ Yarn }) {
-  const EXPECTED_VERSION = "12.0.0-beta.1";
+  const EXPECTED_VERSION = "12.0.0-beta.2";
 
   for (const workspace of Yarn.workspaces()) {
     const { ident } = workspace.pkg;